added modal to check whether they want to delete account

+++ client/src/pages/Settings.js
@@ -1,4 +1,5 @@
 import React, { useEffect, useRef, useState } from "react";
+import OutsideClickHandler from 'react-outside-click-handler';
 
 import API from '../utils/API';
 import axios from 'axios';
@@ -19,6 +20,8 @@ function Settings() {
   // const [newGithub, setNewGithub] = useState(state.user._json.login || "");
   // const [newLanguages, setNewLanguages] = useState(state.languages);
   const [newLocation, setNewLocation] = useState(state.location);
+  
+  const [showModal, setShowModal] = React.useState(false);
   
   const [errorMsgProfile, setErrorMsgProfile] = useState("");
   const [errorMsgPassword, setErrorMsgPassword] = useState("");
@@ -335,12 +338,75 @@ function Settings() {
         <button 
           className="del-account" 
           style={{ width:"200px",backgroundColor:"mediumvioletred", color:"white", borderRadius:"99px"}}
-          onClick={() => {
-            deleteUser();
+          onClick={(e)=>{
+            setShowModal(true)
           }}>
           <strong>Delete Account</strong>
         </button>
       </div>
+
+
+      {showModal ? (
+        <>
+        
+          <div
+            className="justify-center items-center flex overflow-x-hidden overflow-y-auto fixed inset-0 z-50 outline-none focus:outline-none"
+          >
+            <OutsideClickHandler onOutsideClick={() => {
+          console.log('clicking outisde')
+          setShowModal(false)
+          }}> 
+            <div className="relative w-auto my-6 mx-auto max-w-3xl">
+              {/*content*/}
+              <div className="border-0 rounded-lg shadow-lg relative flex flex-col w-full bg-white outline-none focus:outline-none">
+                {/*header*/}
+                {/* <div className="flex items-start justify-between p-5 border-b border-solid border-blueGray-200 rounded-t">
+                  <h3 className="text-3xl font-semibold">
+                    Modal Title
+                  </h3>
+                  <button
+                    className="p-1 ml-auto bg-transparent border-0 text-black opacity-5 float-right text-3xl leading-none font-semibold outline-none focus:outline-none"
+                    onClick={() => setShowModal(false)}
+                  >
+                    <span className="bg-transparent text-black opacity-5 h-6 w-6 text-2xl block outline-none focus:outline-none">
+                      ×
+                    </span>
+                  </button>
+                </div> */}
+                {/*body*/}
+                <div className="relative p-6 flex-auto">
+                  <p className="my-4 text-blueGray-500 text-lg leading-relaxed">
+                    Are you sure you want to delete your account?
+                  </p>
+                </div>
+                {/*footer*/}
+                <div className="flex items-center justify-end rounded-b">
+                  <button
+                    className="text-red-500 background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                    type="button"
+                    onClick={() => {
+                      deleteUser();
+                      setShowModal(false)
+                    }}
+                  >
+                    Yes
+                  </button>
+                  <button
+                    className="background-transparent font-bold uppercase px-6 py-2 text-sm outline-none focus:outline-none mr-1 mb-1 ease-linear transition-all duration-150"
+                    type="button"
+                    onClick={() => setShowModal(false)}
+                  >
+                    No
+                  </button>
+                </div>
+              </div>
+            </div>
+      </OutsideClickHandler>
+          </div>
+          <div className="opacity-25 fixed inset-0 z-40 bg-black"></div>
+          
+        </>
+      ) : null}
     </div>
   );
 }
